Split Cube and Face classes into different files
